requirements.txt
python-casacore
PySide6
overrides
matplotlib
pandas